fix(redteam): enforce dynamic prompt minima
